Create a new .utility DNS zone
This zone will allow to get fixed names to access utility services, such
as the IPMI.
Using names instead of IPs ensures we can easily inject IPv6 in the
game, without having to chase down IPs in the code: just update the DNS
record to get either IPv4, or IPv6, or even dual-stack.

This patch also updates sushy_emulator role to remove some hacks
injected as "quick workaround". The cifmw_sushy_emulator_listen_ip
parameter is now passed as-is and, if needed, any IPv6 we really pass
there would be properly wrapped.
In the reproducer context, it will get a hostname instead, ensuring we
pass via the .utility zone and appropriate records.